dmtxwrite: document new encodation default
Commit da5fec1 ("Changed default mode to optimize best") changed the
default encodation scheme from the previous default of "ASCII" to the
automatic optimizer. However, the output of --help and the manpage still
claim that ASCII is the default, which can cause confusion.

I guess that change also means that this mode is no longer considered
"beta".

Update the documentation accordingly.